session/4 Delegateパターン
Weatherクラスを別ファイルに切り出し
WeatherManagerクラスを作成し、ViewController側からWeatherManagerを呼び出し→WeatherManager側でAPIを利用して天気を取得→取得が完了した段階で呼び出し元のViewControllerにdelegateで処理を委譲するように修正
